[Feature] Add endpoints for visit, visitors and popular

+++ app/api_1_0/url.py
@@ -4,13 +4,13 @@ import string
 import dotenv
 from app.models import LongUrl, ShortUrl, User, Visitor
 from flask import abort, g, jsonify, request, url_for
-from sqlalchemy import desc
+from sqlalchemy import desc, func
 
 from . import api
 from .. import db
 from .authentication import auth
 from .errors import forbidden, not_found, unauthorized
-from .validators import ValidateLongUrl
+from .validators import ValidateLongUrl, ValidateShortUrl
 
 dotenv.load()
 
@@ -122,7 +122,8 @@ def toggle_url_activation(id, activate_or_deactivate):
     if not short_url or short_url.deleted:
         return not_found('No url with the id {}'.format(id))
     if short_url.user == g.current_user:
-        short_url.activate = endpoint_map[activate_or_deactivate]
+        short_url.active = endpoint_map[activate_or_deactivate]
+        db.session.commit()
         return jsonify({'success': True, 'message': url_for(
             'api.shorturls', _external=True) + str(short_url.short_url_id)})
     return unauthorized('Invalid credentials')
@@ -179,6 +180,11 @@ def change_short_url_target(id, url):
     if not long_url_input.validate():
         return forbidden(long_url_input.errors)
     new_long_url = request.json.get('long_url')
+    has_short_url = ShortUrl.query.filter_by(
+        user=g.current_user).filter(ShortUrl.long_url.has(long_url=new_long_url)).first()
+    if has_short_url:
+        return forbidden("The url already has a shortened url '{}'".format(
+            has_short_url.short_url))
     count = len(ShortUrl.query.filter_by(long_url=url.long_url).all())
     if count <= 1:
         LongUrl.query.filter_by(long_url=url.long_url.long_url).delete()
@@ -189,3 +195,87 @@ def change_short_url_target(id, url):
     url.long_url = new_long_url_check
     db.session.commit()
     return jsonify({'success': True, 'message': 'updated'})
+
+
+@api.route('/shorturl/popular/', methods=['GET'])
+def popular():
+    urls = ShortUrl.query(func.count(ShortUrl.visitors.visitor_id).label(
+        'total')). order_by('total desc').all()  # Might consider limits later
+    if not urls:
+        popular_urls = []
+        for url in urls:
+            visitors = Visitor.query(Visitor.visitor_id).filter(
+                ShortUrl.visitors.any(short_url_id=url.short_url_id)).all()
+            popular_urls.append({
+                'short_url_url': url_for(
+                    'api.shorturl', _external=True) + str(url.short_url_id),
+                'short_url': url.short_url,
+                'number_of_visits': url.total,
+                'visitors': [
+                    url_for('api.visitor', vid=visitor.visitor_id,
+                            id=url.short_url_id,
+                            _external=True) for visitor in visitors]
+            })
+        return jsonify({'success': True, 'popular_urls': popular_urls})
+    return not_found('No url was found')
+
+
+@api.route('/shorturl/<int:id>/visitors/', methods=['GET'])
+def visitors(id):
+    if ShortUrl.query.filter_by(
+            short_url_id=id).filter_by(user=g.current_user).first():
+        visitors = Visitor.query(Visitor.visitor_id).filter(
+            ShortUrl.visitors.any(short_url_id=id)).all()
+        if not visitors:
+            visitors_details = []
+            for visitor in visitors:
+                visitor_details = visitor.get_details()
+                visitor_details['visitor_url'] = url_for(
+                    'api.visitor', vid=visitor.visitor_id,
+                    id=id, _external=True)
+                visitors_details.append(visitor_details)
+            return jsonify({'success': True, 'visitors': visitors_details})
+        return not_found('No visitor found for this URL')
+    return not_found("No URL found with id '{}'".format(id))
+
+
+# Abstract visitors detail (keep it DRY)
+@api.route('/shorturl/<int:id>/visitors/<int:vid>', methods=['GET'])
+def visitor(id, vid):
+    if ShortUrl.query.filter_by(
+            short_url_id=id).filter_by(user=g.current_user).first():
+        visitor = Visitor.query.filter_by(visitor_id=vid).first()
+        if visitor:
+            visitor_details = visitor.get_details()
+            visitor_details['visitor_url'] = url_for(
+                'api.visitor', vid=vid, id=id, _external=True)
+            return jsonify({'success': True, 'visitor': visitor_details})
+        return not_found("No visitor to this URL with id '{}'".format(vid))
+    return not_found("No URL found with id '{}'".format(id))
+
+
+# check for deleted and deactivated urls in all other route implementation
+@api.route('/visit/', methods=['POST'])
+def visit():
+    short_url_input = ValidateShortUrl(request)
+    if not short_url_input.validate():
+        return forbidden(short_url_input.errors)
+    short_url = request.json.get('short_url')
+    short_url_details = ShortUrl.query.filter_by(
+        short_url=short_url).filter_by(deleted=0).first()
+    if not short_url_details:
+        return not_found('No matching URL found')
+    elif not short_url_details.active:
+        return forbidden('URL has been deactivated')
+    else:
+        add_visit(short_url_details)
+        return jsonify(
+            {'success': True, 'long_url': short_url_details.long_url.long_url})
+
+
+def add_visit(url):
+    agent = request.user_agent
+    visitor = Visitor(ip_address=request.remote_addr,
+                      browser=agent.browser, platform=agent.platform)
+    visitor.short_urls.append(url)
+    visitor.save()
